### 2018-12-18 #1134 @whedon
[PRE REVIEW]: PyWavelets: A Python package for wavelet analysis
**Submitting author:** @grlee77 (<a href="http://orcid.org/0000-0001-8895-2740">Gregory Lee</a>)
**Repository:** <a href="https://github.com/PyWavelets/pywt" target ="_blank">https://github.com/PyWavelets/pywt</a>
**Version:** v1.0.1
