### 2019-01-11 c8d1204
Fix potential coefficient shape mismatch in WaveletPacket reconstruction.
inverse transforms may produce excess coefficients for some sizes and boundary
modes. these need to be trimmed at the Node level as implemented here.
This commit fixes PyWavelets/pywt#447

Add additional wavelet packet reconstruction tests

remove coefficient trimming from WaveletPacket classes

trimming is now done at the Node level, so this is redundant.

rename data_shape->_data_shape (private)
