+++ pywt/_wavelet_packets.py
@@ -68,6 +68,12 @@ class BaseNode(object):
 
         # data - signal on level 0, coeffs on higher levels
         self.data = data
+        # Need to retain original data size/shape so we can trim any excess
+        # boundary coefficients from the inverse transform.
+        if self.data is None:
+            self._data_shape = None
+        else:
+            self._data_shape = np.asarray(data).shape
 
         self._init_subnodes()
 
@@ -436,6 +442,9 @@ class Node(BaseNode):
                              " from subnodes.")
         else:
             rec = idwt(data_a, data_d, self.wavelet, self.mode)
+            if self._data_shape is not None and (
+                    rec.shape != self._data_shape):
+                rec = rec[tuple([slice(sz) for sz in self._data_shape])]
             if update:
                 self.data = rec
             return rec
@@ -504,6 +513,9 @@ class Node2D(BaseNode):
         else:
             coeffs = data_ll, (data_hl, data_lh, data_hh)
             rec = idwt2(coeffs, self.wavelet, self.mode)
+            if self._data_shape is not None and (
+                    rec.shape != self._data_shape):
+                rec = rec[tuple([slice(sz) for sz in self._data_shape])]
             if update:
                 self.data = rec
             return rec
@@ -568,8 +580,6 @@ class WaveletPacket(Node):
         """
         if self.has_any_subnode:
             data = super(WaveletPacket, self).reconstruct(update)
-            if self.data_size is not None and len(data) > self.data_size:
-                data = data[:self.data_size]
             if update:
                 self.data = data
             return data
@@ -669,8 +679,6 @@ class WaveletPacket2D(Node2D):
         """
         if self.has_any_subnode:
             data = super(WaveletPacket2D, self).reconstruct(update)
-            if self.data_size is not None and (data.shape != self.data_size):
-                data = data[:self.data_size[0], :self.data_size[1]]
             if update:
                 self.data = data
             return data
