+++ pywt/tests/test_wp2d.py
@@ -168,5 +168,14 @@ def test_wavelet_packet_dtypes():
         assert_allclose(r, x, atol=1e-5, rtol=1e-5)
 
 
+def test_2d_roundtrip():
+    # test case corresponding to PyWavelets issue 447
+    original = pywt.data.camera()
+    wp = pywt.WaveletPacket2D(data=original, wavelet='db3', mode='smooth',
+                              maxlevel=3)
+    r = wp.reconstruct()
+    assert_allclose(original, r, atol=1e-12, rtol=1e-12)
+
+
 if __name__ == '__main__':
     run_module_suite()
