### 2019-01-11 d31ef2b
Fix potential coefficient shape mismatch in WaveletPacket
inverse transforms may produce excess coefficients for some sizes and boundary
modes. these need to be trimmed at the Node level as implemented here.
This commit fixes PyWavelets/pywt#447
