+++ pywt/_wavelet_packets.py
@@ -68,6 +68,12 @@ class BaseNode(object):
 
         # data - signal on level 0, coeffs on higher levels
         self.data = data
+        # Need to retain original data size/shape so we can trim any excess
+        # boundary coefficients from the inverse transform.
+        if self.data is None:
+            self.data_size = None
+        else:
+            self.data_size = np.asarray(data).shape
 
         self._init_subnodes()
 
@@ -436,6 +442,8 @@ class Node(BaseNode):
                              " from subnodes.")
         else:
             rec = idwt(data_a, data_d, self.wavelet, self.mode)
+            if self.data_size is not None and (rec.shape != self.data_size):
+                rec = rec[tuple([slice(sz) for sz in self.data_size])]
             if update:
                 self.data = rec
             return rec
@@ -504,6 +512,8 @@ class Node2D(BaseNode):
         else:
             coeffs = data_ll, (data_hl, data_lh, data_hh)
             rec = idwt2(coeffs, self.wavelet, self.mode)
+            if self.data_size is not None and (rec.shape != self.data_size):
+                rec = rec[tuple([slice(sz) for sz in self.data_size])]
             if update:
                 self.data = rec
             return rec
