### 2019-01-11 e04ee83
remove coefficient trimming from WaveletPacket classes
trimming is now done at the Node level, so this is redundant.
