### 2019-01-19 #1134 @labarba
I wonder if @rafat would be willing to help us with this review?

👋 @rafat —Are you familiar with the Journal of Open Source Software (JOSS)? JOSS is a free, open-source, community driven and developer-friendly online journal. It aims to bridge the gap of research software receiving citation credit. JOSS reviews involve downloading and installing the software, and inspecting the repository for key elements. Have a look at the JOSS [Review Process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
