### 2019-01-20 #1134 @jedbrown
Thanks @labarba for catching this stall and thanks @rafat for being available to review.

@grlee77 There appears to be a missing reference ("manuscript by Stephan Mallat (???)") in your paper. Also, JOSS papers will include a Zenodo archive of the repository (post-review), so no need to cite an existing archive in the text.  Additionally, would it be appropriate to mention related work such as [Kymatio](https://github.com/kymatio/kymatio)?
