### 2019-01-31 e1a33c2
handle mixed dtype cofficients correctly across inverse transforms
use _check_dtype on the coarsest level approximation coeffs

output array dtype is upcast if any detail coefficients have higher precision
than the approximation coefficients.

fix: make dtype upcast in idwt respect complex dtypes

add a dtype upcast to idwtn to match the behavior of idwt

test inverse SWT functions with mixed dtype coefficients

test idwt and idwtn with mixed complex dtypes as well

test mixed precision with wavedec/waverec as well

idwtn: only compare dtypes if both coeffs are not None
