### 2019-01-31 5fdd6bd
use _check_dtype on the coarsest level approximation coeffs
output array dtype is upcast if any detail coefficients have higher precision
than the approximation coefficients.
