### 2019-01-31 afa2eb1
add a dtype upcast to idwtn to match the behavior of idwt
