### 2019-01-31 cc6ea9d
test mixed precision with wavedec/waverec as well
