### 2019-01-31 fb0fa8c
idwtn: only compare dtypes if both coeffs are not None
