### 2019-02-04 9dc4496
BLD: pin cython language level to '2'
This is a simpler alternative to https://github.com/PyWavelets/pywt/pull/435
