### 2019-02-04 cf492ae
move language_level Cython setting to setup.py
as suggested by tacaswell in gh-451
