+++ pywt/_extensions/_swt.pyx
@@ -1,4 +1,4 @@
-#cython: language_level=3, boundscheck=False, wraparound=False
+#cython: boundscheck=False, wraparound=False
 from . cimport common
 from . cimport c_wt
 
