+++ pywt/_extensions/_pywt.pyx
@@ -1,4 +1,3 @@
-#cython: language_level=3
 # Copyright (c) 2006-2012 Filip Wasilewski <http://en.ig.ma/>
 # Copyright (c) 2012-2018 The PyWavelets Developers
 #                         <https://github.com/PyWavelets/pywt>
