### 2019-02-04 1b2fb50
remove pin from Cython version used by readthedocs
use the minimum supported Cython version in one of the TravisCI runs

bumpy minimum Cython version required on RTD
