### 2019-02-04 c414c63
use the minimum supported Cython version in one of the TravisCI runs
