+++ .travis.yml
@@ -16,18 +16,21 @@ matrix:
       env:
         - NUMPYSPEC=numpy
         - MATPLOTLIBSPEC=matplotlib
+        - CYTHONSPEC=cython
         - USE_WHEEL=1
     - os: linux
       python: 3.7-dev
       env:
         - NUMPYSPEC=numpy
         - MATPLOTLIBSPEC=matplotlib
+        - CYTHONSPEC=cython
         - USE_SDIST=1
     - os: linux
       python: 3.5
       env:
         - NUMPYSPEC="numpy==1.13.3"
         - MATPLOTLIBSPEC=matplotlib
+        - CYTHONSPEC="cython==0.23.5"
         - REFGUIDE_CHECK=1  # run doctests only
     - os: osx
       osx_image: xcode7.3
@@ -35,6 +38,7 @@ matrix:
       env:
         - NUMPYSPEC=numpy
         - MATPLOTLIBSPEC=matplotlib
+        - CYTHONSPEC=cython
         - TRAVIS_PYTHON_VERSION=3.5
 
 cache: pip
@@ -51,7 +55,8 @@ before_install:
   # Set numpy version first, other packages link against it
   - pip install $NUMPYSPEC
   - pip install $MATPLOTLIBSPEC
-  - pip install Cython nose coverage codecov futures
+  - pip install $CYTHONSPEC
+  - pip install nose coverage codecov futures
   - set -o pipefail
   - if [ "${USE_WHEEL}" == "1" ]; then pip install wheel; fi
   - |
