### 2019-02-04 288ccb3
update the Cython version requirement in the docs [skip ci][ci skip]
