### 2019-02-05 #1134 @grlee77
Apologies for missing the notification above regarding the reference.  It should be fixed now.

> Additionally, would it be appropriate to mention related work such as [Kymatio](https://github.com/kymatio/kymatio)?

Yeah, I saw this library recently and can add a mention to the manuscript.  Kymatio implements a scattering transform that I haven't studied in detail, but appears to be useful for things like classification. I believe the underlying wavelets involved are implemented in the Fourier domain whereas everything in PyWavelets is done via convolution with FIR filters.  PyWavelets has a wide range of classical wavelet transforms with a well defined inverse transform while I think it may be possible to invert a scattering transform, but it is not so straightforward.
