### 2019-02-05 #1237 @whedon
[REVIEW]: PyWavelets: A Python package for wavelet analysis
**Submitting author:** @grlee77 (<a href="http://orcid.org/0000-0001-8895-2740">Gregory Lee</a>)
**Repository:** <a href="https://github.com/PyWavelets/pywt" target ="_blank">https://github.com/PyWavelets/pywt</a>
**Version:** v1.0.3
