### 2019-02-05 #1237 @jedbrown
@rafat @souopgui :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the PyWavelets repository). I'll be watching this thread if you have any questions.

@grlee77 Can you clarify whether this review should take place on the 'master' branch (where your paper sources currently reside) or in the '1.0.x' branch which contains the stated version? Thanks.
