### 2019-02-07 #1237 @grlee77
The features and functionality are the same in both 1.0.1 and master so I am fine with reviewing based on 1.0.1. The main difference is that master has dropped Python 2 support and has a few minor fixes. We will eventually include the fixes in a future release from the 1.0.x branch as well.
