### 2019-02-07 #1237 @grlee77
There are a number of basic demos of usage in the `demo` folder of the repository. The API documentation is pretty complete, but the docs don't have a lot of tutorials/demos. Some concrete examples for scientific applications are provided below.

There are a some image denoising demos functions in scikit-image that rely on PyWavelets (specifically for `skimage.restoration.estimate_sigma` and `skimage.restoration.denoise_wavelet`):
http://scikit-image.org/docs/dev/auto_examples/filters/plot_cycle_spinning.html
http://scikit-image.org/docs/dev/auto_examples/filters/plot_denoise_wavelet.html

An example of wrapping the wavelet transform into an operator (e.g. for potential use as a sparsity-enforcing regularizer in iterative medical image reconstruction) is available within ODL:
https://github.com/odlgroup/odl/blob/master/examples/trafos/wavelet_trafo.py

For time series analysis, a user working in the field of geophysics recently posted the following notebook with some examples to the mailing list:
https://gist.github.com/aprovecharLab/09ded80ae59ea510334e00926a15920f
