### 2019-02-07 #1237 @jedbrown
Thanks, @grlee77. Note that upon acceptance, we'll ask you to tag and archive a release that includes any revisions that may have resulted from the review. That archive should include the final sources for the paper, which are currently in 'master'.
