### 2019-02-07 #1237 @grlee77
Okay, in that case, please just review from master and we will tag a new release once the review is completed.  The 1.0.x branch was created to allow maintenance/bug fixes releases for Python 2.7 support through 2020 or so, but any new features introduced will not be backported to that branch.
