### 2019-02-11 afec3e4
rename data_shape->_data_shape (private)
