### 2019-02-14 de2ee36
fix iswtn axis-specific transform bug
