### 2019-02-14 a17fedf
TST: add tests for round-trip swtn/iswtn with non-uniform shape
