### 2019-02-15 #1237 @rafat
@grlee77 There are a few  minor issues with .rst files n the doc/source folder that are causing test_doc.py to fail when I run it manually from the folder. For example, line 335 of expected value has "Short name:     db" but the code correctly returns "Short name:     gaus" .  There are no issues with the Python code as far as can I tell but you may want to take a look at the .rst files.
