### 2019-02-15 0cb2ba0
add numpy 1.13 legacy print option to test_doc.py
