### 2019-02-15 7109cbe
fix remaining doctest failures (as found by test_doc.py)
