### 2019-02-15 4ae270b
run test_doc.py on Travis
nosetests skips the tests in this file, so run it separately.
