### 2019-02-15 #1237 @grlee77
Thanks for pointing that out. It seems the tests in that file were being skipped by our CI scripts. I have fixed the failures and made sure TravisCI is running these tests in PyWavelets/pywt#463.
