### 2019-02-17 #1237 @souopgui
@grlee77 Quick questions to help me finalize my review:

1. Apart from the references to Mallat "A Theory for Multiresolution Signal Decomposition: The Wavelet Representation", Wickerhauser "Adapted Wavelet Analysis from Theory to Software" and the thesis that started this work, is there other part of the doc that I am missing? I think that for a topic like the wavelet, there should be a little more documentations, think of Haar, Morlet, Meyer, coifman, Daubechies, etc. ?
2. Is there a way to try the automated test outside of the install procedure?
