### 2019-02-18 #1237 @grlee77
> 1. is there other part of the doc that I am missing?

I think you are asking about the PyWavelets documentation and not the manuscript itself? Specific functions do have additional references to the literature (see for example the docs for [threshold](https://pywavelets.readthedocs.io/en/latest/ref/thresholding-functions.html#pywt.threshold) or [fswavedecn](https://pywavelets.readthedocs.io/en/latest/ref/nd-dwt-and-idwt.html#multilevel-fully-separable-decomposition-fswavedecn), but I agree that we are lacking an overview page with basic theory and key references, etc. Wavelets are well covered in many textbooks at this point, but it might still be worth having an abbreviated summary in the PyWavelet docs.

> 2. Is there a way to try the automated test outside of the install procedure?

Yes, try:
```python
import pywt
pywt.test()
```
