### 2019-02-26 f597b43
add missing CYTHONSPEC for python 2.7
