+++ setup.py
@@ -15,7 +15,7 @@ from setuptools import setup, Extension
 MAJOR = 1
 MINOR = 0
 MICRO = 2
-ISRELEASED = False
+ISRELEASED = True
 VERSION = '%d.%d.%d' % (MAJOR, MINOR, MICRO)
 
 
