### 2019-02-27 fe1913a
specify output_format explicitly in the unravel_coeffs example
