+++ pywt/_multilevel.py
@@ -1120,7 +1120,8 @@ def unravel_coeffs(arr, coeff_slices, coeff_shapes, output_format='wavedecn'):
     >>> cam = pywt.data.camera()
     >>> coeffs = pywt.wavedecn(cam, wavelet='db2', level=3)
     >>> arr, coeff_slices, coeff_shapes = pywt.ravel_coeffs(coeffs)
-    >>> coeffs_from_arr = pywt.unravel_coeffs(arr, coeff_slices, coeff_shapes)
+    >>> coeffs_from_arr = pywt.unravel_coeffs(arr, coeff_slices, coeff_shapes,
+    >>>                                       output_format='wavedecn')
     >>> cam_recon = pywt.waverecn(coeffs_from_arr, wavelet='db2')
     >>> assert_array_almost_equal(cam, cam_recon)
 
