### 2019-02-27 088e1b9
add user-friendly error messages related to coefficient format
