### 2019-02-27 79ebf0e
remove outdated AttributeError check.
This case can never be reached due to the new ValueError check above
