### 2019-02-27 c90cf85
fix docstring example
