### 2019-03-13 a1d484f
add Kymatio to the discussion
