### 2019-03-13 #1237 @grlee77
hey @rafat, just checking if you are still working on this and if there are any other items left to address?

I just opened a PR at PyWavelets/pywt#475 to add a reference to the book by Daubechies, more clearly link Mallat to the multiresolution property of the transform, and to add a brief discussion of Kymatio in relation to PyWavelets.
