### 2019-03-13 #1237 @rafat
Hi @grlee77 @jedbrown I finished my review a couple of weeks ago. Pywavelets is a very good software and everything works as it should on two different systems (Windows and Linux).
