### 2019-03-15 297b1df
add trim_approx boolean to swt cython routines
If trim_approx is set to True, we will retain approximation coefficients
only for the final level of decomposition as is currently done for the
DWT functions.

add trim_approx option to all SWT transforms
