+++ pywt/_swt.py
@@ -14,7 +14,8 @@ from ._utils import _as_wavelet, _wavelets_per_axis
 __all__ = ["swt", "swt_max_level", 'iswt', 'swt2', 'iswt2', 'swtn', 'iswtn']
 
 
-def swt(data, wavelet, level=None, start_level=0, axis=-1):
+def swt(data, wavelet, level=None, start_level=0, axis=-1,
+        trim_approx=False):
     """
     Multilevel 1D stationary wavelet transform.
 
@@ -33,6 +34,8 @@ def swt(data, wavelet, level=None, start_level=0, axis=-1):
     axis: int, optional
         Axis over which to compute the SWT. If not given, the
         last axis is used.
+    trim_approx : bool, optional
+        If True, approximation coefficients at the final level are retained.
 
     Returns
     -------
@@ -49,6 +52,13 @@ def swt(data, wavelet, level=None, start_level=0, axis=-1):
 
             [(cAm+n, cDm+n), ..., (cAm+1, cDm+1), (cAm, cDm)]
 
+        If ``trim_approx`` is ``True``, then the output list is exactly as in
+        ``pywt.wavedec``, where the first coefficient in the list is the
+        approximation coefficient at the final level and the rest are the
+        detail coefficients::
+
+            [cAn, cDn, ..., cD2, cD1]
+
     Notes
     -----
     The implementation here follows the "algorithm a-trous" and requires that
@@ -58,11 +68,15 @@ def swt(data, wavelet, level=None, start_level=0, axis=-1):
     """
     if not _have_c99_complex and np.iscomplexobj(data):
         data = np.asarray(data)
-        coeffs_real = swt(data.real, wavelet, level, start_level)
-        coeffs_imag = swt(data.imag, wavelet, level, start_level)
-        coeffs_cplx = []
-        for (cA_r, cD_r), (cA_i, cD_i) in zip(coeffs_real, coeffs_imag):
-            coeffs_cplx.append((cA_r + 1j*cA_i, cD_r + 1j*cD_i))
+        coeffs_real = swt(data.real, wavelet, level, start_level, trim_approx)
+        coeffs_imag = swt(data.imag, wavelet, level, start_level, trim_approx)
+        if not trim_approx:
+            coeffs_cplx = []
+            for (cA_r, cD_r), (cA_i, cD_i) in zip(coeffs_real, coeffs_imag):
+                coeffs_cplx.append((cA_r + 1j*cA_i, cD_r + 1j*cD_i))
+        else:
+            coeffs_cplx = [cr + 1j*ci
+                           for (cr, ci) in zip(coeffs_real, coeffs_imag)]
         return coeffs_cplx
 
     # accept array_like input; make a copy to ensure a contiguous array
@@ -80,10 +94,10 @@ def swt(data, wavelet, level=None, start_level=0, axis=-1):
         level = swt_max_level(data.shape[axis])
 
     if data.ndim == 1:
-        ret = _swt(data, wavelet, level, start_level)
+        ret = _swt(data, wavelet, level, start_level, trim_approx)
     else:
-        ret = _swt_axis(data, wavelet, level, start_level, axis)
-    return [(np.asarray(cA), np.asarray(cD)) for cA, cD in ret]
+        ret = _swt_axis(data, wavelet, level, start_level, axis, trim_approx)
+    return ret
 
 
 def iswt(coeffs, wavelet):
@@ -114,13 +128,26 @@ def iswt(coeffs, wavelet):
     array([ 1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.])
     """
     # copy to avoid modification of input data
-    dt = _check_dtype(coeffs[0][0])
[…]
