### 2019-03-15 8dec196
add test cases for swt transforms with trim_approx=True
