### 2019-03-15 970643b
ENH: add normalization option to swt functions
TST: verify perfect round trip transform when norm=True as well
