+++ pywt/tests/test_swt.py
@@ -153,11 +153,12 @@ def test_swt_iswt_integration():
             current_wavelet.rec_len))))
         input_length = 2**(input_length_power + max_level - 1)
         X = np.arange(input_length)
-        for trim_approx in [True, False]:
-            coeffs = pywt.swt(X, current_wavelet, max_level,
-                              trim_approx=trim_approx)
-            Y = pywt.iswt(coeffs, current_wavelet)
-            assert_allclose(Y, X, rtol=1e-5, atol=1e-7)
+        for norm in [True, False]:
+            for trim_approx in [True, False]:
+                coeffs = pywt.swt(X, current_wavelet, max_level,
+                                  trim_approx=trim_approx, norm=norm)
+                Y = pywt.iswt(coeffs, current_wavelet, norm=norm)
+                assert_allclose(Y, X, rtol=1e-5, atol=1e-7)
 
 
 def test_swt_dtypes():
@@ -237,11 +238,12 @@ def test_swt2_iswt2_integration(wavelets=None):
         input_length = 2**(input_length_power + max_level - 1)
         X = np.arange(input_length**2).reshape(input_length, input_length)
 
-        for trim_approx in [True, False]:
-            coeffs = pywt.swt2(X, current_wavelet, max_level,
-                               trim_approx=trim_approx)
-            Y = pywt.iswt2(coeffs, current_wavelet)
-            assert_allclose(Y, X, rtol=1e-5, atol=1e-5)
+        for norm in [True, False]:
+            for trim_approx in [True, False]:
+                coeffs = pywt.swt2(X, current_wavelet, max_level,
+                                   trim_approx=trim_approx, norm=norm)
+                Y = pywt.iswt2(coeffs, current_wavelet, norm=norm)
+                assert_allclose(Y, X, rtol=1e-5, atol=1e-5)
 
 
 def test_swt2_iswt2_quick():
@@ -359,12 +361,13 @@ def test_swtn_iswtn_integration(wavelets=None):
                 N = 2**(input_length_power + max_level - 1)
                 X = np.arange(N**ndim).reshape((N, )*ndim)
 
-                for trim_approx in [True, False]:
-                    coeffs = pywt.swtn(X, wav, max_level, axes=axes,
-                                       trim_approx=trim_approx)
-                    coeffs_copy = deepcopy(coeffs)
-                    Y = pywt.iswtn(coeffs, wav, axes=axes)
-                    assert_allclose(Y, X, rtol=1e-5, atol=1e-5)
+                for norm in [True, False]:
+                    for trim_approx in [True, False]:
+                        coeffs = pywt.swtn(X, wav, max_level, axes=axes,
+                                           trim_approx=trim_approx, norm=norm)
+                        coeffs_copy = deepcopy(coeffs)
+                        Y = pywt.iswtn(coeffs, wav, axes=axes, norm=norm)
+                        assert_allclose(Y, X, rtol=1e-5, atol=1e-5)
 
                 # verify the inverse transform didn't modify any coeffs
                 for c, c2 in zip(coeffs, coeffs_copy):
