+++ pywt/_swt.py
@@ -6,7 +6,7 @@ import numpy as np
 from ._c99_config import _have_c99_complex
 from ._extensions._dwt import idwt_single
 from ._extensions._swt import swt_max_level, swt as _swt, swt_axis as _swt_axis
-from ._extensions._pywt import Modes, _check_dtype
+from ._extensions._pywt import Wavelet, Modes, _check_dtype
 from ._multidim import idwt2, idwtn
 from ._utils import _as_wavelet, _wavelets_per_axis
 
@@ -14,8 +14,13 @@ from ._utils import _as_wavelet, _wavelets_per_axis
 __all__ = ["swt", "swt_max_level", 'iswt', 'swt2', 'iswt2', 'swtn', 'iswtn']
 
 
+def _rescale_wavelet_filterbank(wavelet, sf):
+    return Wavelet(wavelet.name + 'r',
+                   [np.asarray(f) * sf for f in wavelet.filter_bank])
+
+
 def swt(data, wavelet, level=None, start_level=0, axis=-1,
-        trim_approx=False):
+        trim_approx=False, norm=False):
     """
     Multilevel 1D stationary wavelet transform.
 
@@ -36,6 +41,11 @@ def swt(data, wavelet, level=None, start_level=0, axis=-1,
         last axis is used.
     trim_approx : bool, optional
         If True, approximation coefficients at the final level are retained.
+    norm : bool, optional
+        If True, transform is normalized so that the energy of the coefficients
+        will be equal to the energy of ``data``. In other words,
+        ``np.linalg.norm(data.ravel())`` will equal the norm of the
+        concatenated transform coefficients when ``trim_approx`` is True.
 
     Returns
     -------
@@ -65,6 +75,23 @@ def swt(data, wavelet, level=None, start_level=0, axis=-1,
     the signal length along the transformed axis be a multiple of ``2**level``.
     If this is not the case, the user should pad up to an appropriate size
     using a function such as ``numpy.pad``.
+
+    A primary benefit of this transform in comparison to its decimated
+    counterpart (``pywt.wavedecn``), is that it is shift-invariant. This comes
+    at cost of redundancy in the transform (the size of the output coefficients
+    is larger than the input).
+
+    When the following three conditions are true::
+
+        1.) The wavelet is orthogonal
+        2.) ``swt`` is called with ``norm=True``
+        3.) ``swt`` is called with ``trim_approx=True``
+
+    the transform has the following additional properties that may be
+    desirable in applications:
+        1.) energy is conserved
+        2.) variance is partitioned across scales
+
     """
     if not _have_c99_complex and np.iscomplexobj(data):
         data = np.asarray(data)
@@ -84,6 +111,8 @@ def swt(data, wavelet, level=None, start_level=0, axis=-1,
     data = np.array(data, dtype=dt)
 
     wavelet = _as_wavelet(wavelet)
+    if norm:
+        wavelet = _rescale_wavelet_filterbank(wavelet, 1/np.sqrt(2))
 
     if axis < 0:
         axis = axis + data.ndim
@@ -100,7 +129,7 @@ def swt(data, wavelet, level=None, start_level=0, axis=-1,
     return ret
 
 
-def iswt(coeffs, wavelet):
+def iswt(coeffs, wavelet, norm=False):
     """
     Multilevel 1D inverse discrete stationary wavelet transform.
 
@@ -115,6 +144,10 @@ def iswt(coeffs, wavelet):
         ``start_level`` from ``pywt.swt``.
     wavelet : Wavelet object or name string
         Wavelet to use
+    norm : bool, optional
+        Controls the normalization used by the inverse transform. This must
+        be set equal to the value that was used by ``pywt.swt`` to preserve the
+        energy of a round-trip transform.
 
     Returns
     -------
[…]
