### 2019-03-15 31b4262
TST: test properties of the swt when norm=True
1.) energy preservation
2.) variance partition
