### 2019-03-15 5776445
add Userwarning for swt with norm=True when wavelet is not orthogonal
