### 2019-03-15 179896b
test pywt.ravel_coeffs and pywt.unravel_coeffs with swt functions
