### 2019-03-15 6ac0fa3
update docstring of ravel_coeffs and unravel_coeffs
allow output formats of 'swt', 'swt2' and 'swtn' (equivalent to 'wavedec', 'wavedec2', 'wavedecn')
