+++ pywt/tests/test_swt.py
@@ -622,9 +622,9 @@ def test_swtn_variance_and_energy_preservation():
 def test_swt_ravel_and_unravel():
     # When trim_approx=True, all swt functions can user pywt.ravel_coeffs
     for ndim, _swt, _iswt, ravel_type in [
-            (1, pywt.swt, pywt.iswt, 'wavedec'),
-            (2, pywt.swt2, pywt.iswt2, 'wavedec2'),
-            (3, pywt.swtn, pywt.iswtn, 'wavedecn')]:
+            (1, pywt.swt, pywt.iswt, 'swt'),
+            (2, pywt.swt2, pywt.iswt2, 'swt2'),
+            (3, pywt.swtn, pywt.iswtn, 'swtn')]:
         x = np.ones((16, ) * ndim)
         c = _swt(x, 'sym2', level=3, trim_approx=True)
         arr, slices, shapes = pywt.ravel_coeffs(c)
