+++ pywt/_multilevel.py
@@ -1009,7 +1009,9 @@ def ravel_coeffs(coeffs, axes=None):
     ----------
     coeffs : array-like
         A list of multilevel wavelet coefficients as returned by
-        `wavedec`, `wavedec2` or `wavedecn`.
+        `wavedec`, `wavedec2` or `wavedecn`. This function is also compatible
+        with the output of `swt`, `swt2` and `swtn` if those functions were
+        called with `trim_approx=True`.
     axes : sequence of ints, optional
         Axes over which the DWT that created `coeffs` was performed. The
         default value of None corresponds to all axes.
@@ -1093,14 +1095,15 @@ def unravel_coeffs(arr, coeff_slices, coeff_shapes, output_format='wavedecn'):
     arr : array-like
         An array containing all wavelet coefficients. This should have been
         generated by applying `ravel_coeffs` to the output of `wavedec`,
-        `wavedec2` or `wavedecn`.
+        `wavedec2` or `wavedecn` (or via `swt`, `swt2` or `swtn` with
+        `trim_approx=True`).
     coeff_slices : list of tuples
         List of slices corresponding to each coefficient as obtained from
         `ravel_coeffs`.
     coeff_shapes : list of tuples
         List of shapes corresponding to each coefficient as obtained from
         `ravel_coeffs`.
-    output_format : {'wavedec', 'wavedec2', 'wavedecn'}, optional
+    output_format : {'wavedec', 'wavedec2', 'wavedecn', 'swt', 'swt2', 'swtn'}, optional
         Make the form of the unraveled coefficients compatible with this type
         of multilevel transform. The default is 'wavedecn'.
 
@@ -1141,13 +1144,13 @@ def unravel_coeffs(arr, coeff_slices, coeff_shapes, output_format='wavedecn'):
     for n in range(1, len(coeff_slices)):
         slice_dict = coeff_slices[n]
         shape_dict = coeff_shapes[n]
-        if output_format == 'wavedec':
+        if output_format in ['wavedec', 'swt']:
             d = arr[slice_dict['d']].reshape(shape_dict['d'])
-        elif output_format == 'wavedec2':
+        elif output_format in ['wavedec2', 'swt2']:
             d = (arr[slice_dict['da']].reshape(shape_dict['da']),
                  arr[slice_dict['ad']].reshape(shape_dict['ad']),
                  arr[slice_dict['dd']].reshape(shape_dict['dd']))
-        elif output_format == 'wavedecn':
+        elif output_format in ['wavedecn', 'swtn']:
             d = {}
             for k, v in coeff_slices[n].items():
                 d[k] = arr[v].reshape(shape_dict[k])
