### 2019-03-15 e2d7faa
DOC: mention similarity to the MODWT in the swt docstring
