### 2019-03-16 6dbff7c
add functions for mra additive decomposition
