+++ pywt/_mra.py
@@ -1,0 +1,254 @@
+from functools import partial, reduce
+
+import numpy as np
+
+from ._swt import swt, iswt, swt2, iswt2, swtn, iswtn
+from ._multilevel import (wavedec, waverec, wavedec2, waverec2, wavedecn,
+                          waverecn, _prep_axes_wavedecn)
+from ._utils import _wavelets_per_axis, _modes_per_axis
+
+
+__all__ = ["mra", "mra2", "mran"]
+
+
+def mra(data, wavelet, level=None, axis=-1, transform='swt',
+        mode='periodization'):
+
+    if transform == 'swt':
+        if mode != 'periodization':
+            raise ValueError(
+                "transform swt only supports mode='periodization'")
+        kwargs = dict(wavelet=wavelet, norm=True)
+        forward = partial(swt, level=level, trim_approx=True, **kwargs)
+        if axis % data.ndim != data.ndim - 1:
+            raise ValueError("swt only supports axis=-1")
+        inverse = partial(iswt, **kwargs)
+    elif transform == 'dwt':
+        kwargs = dict(wavelet=wavelet, mode=mode, axis=axis)
+        forward = partial(wavedec, level=level, **kwargs)
+        inverse = partial(waverec, **kwargs)
+    else:
+        raise ValueError("unrecognized transform: {}".format(transform))
+
+    wav_coeffs = forward(data)
+
+    mra_coeffs = []
+    nc = len(wav_coeffs)
+    z = np.zeros_like(wav_coeffs[0])
+    tmp = [z, ] * nc
+    for j in range(nc):
+        # tmp has arrays of zeros except for the jth entry
+        tmp[j] = wav_coeffs[j]
+
+        # reconstruct
+        rec = inverse(tmp)
+        if rec.shape != data.shape:
+            # trim any excess coefficients
+            rec = rec[tuple([slice(sz) for sz in data.shape])]
+        mra_coeffs.append(rec)
+
+        # restore zeros
+        tmp[j] = z
+    return mra_coeffs
+
+
+def imra(mra_coeffs):
+    """Inverse 1D multiresolution analysis via summation.
+
+    Parameters
+    ----------
+    mra_coeffs : list of ndarray
+        Multiresolution analysis coefficients as returned by `mra`.
+
+    Returns
+    -------
+    rec : ndarray
+        The reconstructed signal.
+    """
+    return reduce(lambda x, y: x + y, mra_coeffs)
+
+
+def mra2(data, wavelet, level=None, axes=(-2, -1), transform='swt2',
+         mode='periodization'):
+
+    if transform == 'swt2':
+        if mode != 'periodization':
+            raise ValueError(
+                "transform swt only supports mode='periodization'")
+        if axes != (-2, -1):
+            raise ValueError("axes argument not supported for mode swt2")
+        kwargs = dict(wavelet=wavelet, norm=True)
+        forward = partial(swt2, level=level, trim_approx=True, **kwargs)
+        inverse = partial(iswt2, **kwargs)
+    elif transform == 'dwt2':
+        kwargs = dict(wavelet=wavelet, mode=mode, axes=axes)
+        forward = partial(wavedec2, level=level, **kwargs)
+        inverse = partial(waverec2, **kwargs)
+    else:
+        raise ValueError("unrecognized transform: {}".format(transform))
+
+    wav_coeffs = forward(data)
+
+    mra_coeffs = []
+    nc = len(wav_coeffs)
+    z = np.zeros_like(wav_coeffs[0])
+    tmp = [z]
+    for j in range(1, nc):
+        tmp.append([np.zeros_like(c) for c in wav_coeffs[j]])
+
+    # tmp has arrays of zeros except for the jth entry
+    tmp[0] = wav_coeffs[0]
[…]
