### 2019-03-16 365a365
add docstrings
