### 2019-03-16 3cb5436
move pad from _doc_utils.py to _dwt.py and add docstring
