### 2019-03-17 76a94b0
remove misc residual nose-related code
