### 2019-03-17 1ff0132
add n-dimensional support to pad
correct boundary mode name: zeros->zero
