### 2019-03-17 26579c2
TST: add tests for pywt.pad
