### 2019-03-17 41cc925
DOC: add pad to the API documentation
DOC: enable intersphinx extension to allow linking to numpy docs

DOC: consistently link Modes documentation from docstrings
