### 2019-03-17 0020bdf
update appveyor.yml to use pytest
try pip-based install on appveyor
