### 2019-03-27 05fbd25
TST: add PytestTester from numpy
