### 2019-03-27 3358b5d
TST: taylor PytestTester for PyWavelets rather than NumPy
