+++ pywt/_pytesttester.py
@@ -5,13 +5,13 @@ This module implements the ``test()`` function for NumPy modules. The usual
 boiler plate for doing that is to put the following in the module
 ``__init__.py`` file::
 
-    from numpy._pytesttester import PytestTester
+    from pywt._pytesttester import PytestTester
     test = PytestTester(__name__).test
     del PytestTester
 
 
 Warnings filtering and other runtime settings should be dealt with in the
-``pytest.ini`` file in the numpy repo root. The behavior of the test depends on
+``pytest.ini`` file in the pywt repo root. The behavior of the test depends on
 whether or not that file is found as follows:
 
 * ``pytest.ini`` is present (develop mode)
@@ -20,12 +20,9 @@ whether or not that file is found as follows:
     DeprecationWarnings and PendingDeprecationWarnings are ignored, other
     warnings are passed through.
 
-In practice, tests run from the numpy repo are run in develop mode. That
+In practice, tests run from the PyWavelets repo are run in develop mode. That
 includes the standard ``python runtests.py`` invocation.
 
-This module is imported by every numpy subpackage, so lies at the top level to
-simplify circular import issues. For the same reason, it contains no numpy
-imports at module scope, instead importing numpy within function calls.
 """
 from __future__ import division, absolute_import, print_function
 
@@ -35,23 +32,24 @@ import os
 __all__ = ['PytestTester']
 
 
-
-def _show_numpy_info():
-    import numpy as np
-
-    print("NumPy version %s" % np.__version__)
-    relaxed_strides = np.ones((10, 1), order="C").flags.f_contiguous
-    print("NumPy relaxed strides checking option:", relaxed_strides)
+def _show_pywt_info():
+    import pywt
+    from pywt._c99_config import _have_c99_complex
+    print("PyWavelets version %s" % pywt.__version__)
+    if _have_c99_complex:
+        print("Compiled with C99 complex support.")
+    else:
+        print("Compiled without C99 complex support.")
 
 
 class PytestTester(object):
     """
     Pytest test runner.
 
-    This class is made available in ``numpy.testing``, and a test function
+    This class is made available in ``pywt.testing``, and a test function
     is typically added to a package's __init__.py like so::
 
-      from numpy.testing import PytestTester
+      from pywt.testing import PytestTester
       test = PytestTester(__name__).test
       del PytestTester
 
@@ -103,13 +101,6 @@ class PytestTester(object):
         result : bool
             Return True on success, false otherwise.
 
-        Notes
-        -----
-        Each NumPy module exposes `test` in its namespace to run all tests for
-        it. For example, to run all tests for numpy.lib:
-
-        >>> np.lib.test() #doctest: +SKIP
-
         Examples
         --------
         >>> result = np.lib.test() #doctest: +SKIP
@@ -120,14 +111,6 @@ class PytestTester(object):
 
         """
         import pytest
-        import warnings
-
-        #FIXME This is no longer needed? Assume it was for use in tests.
-        # cap verbosity at 3, which is equivalent to the pytest '-vv' option
-        #from . import utils
-        #verbose = min(int(verbose), 3)
-        #utils.verbose = verbose
-        #
 
         module = sys.modules[self.module_name]
         module_path = os.path.abspath(module.__path__[0])
@@ -138,39 +121,12 @@ class PytestTester(object):
         # offset verbosity. The "-q" cancels a "-v".
         pytest_args += ["-q"]
 
-        # Filter out distutils cpu warnings (could be localized to
-        # distutils tests). ASV has problems with top level import,
-        # so fetch module for suppression here.
-        with warnings.catch_warnings():
[…]
