### 2019-04-04 #1237 @grlee77
Just wanted to check in with @souopgui as the review seems to have stalled: Are there still remaining items to address?

I think the only substantial change in master made subsequent to the review comments was updating the underlying testing framework used from `nose` to `pytest` since `nose` is no longer being updated. The tests themselves are the same as before (but hopefully will be refactored to some extent in the future to also make better use of features like `pytest.mark.parametrize`).
