### 2019-04-05 #1237 @souopgui
@grlee77 , I have not paid attention to email coming from github lately.
The toolbox is great and I am already recommending people to give it a try. For a topic of the importance of the wavelet, a toolbox like this one was missing. With the implementation of all the filters you guys implemented, this is a great tools.
With that, I am done with my review. Being less familiar with the review on github, if I am missing something to close the process, please let me know. And sorry a again for being out of the loop for this long.
