### 2019-04-05 #1237 @jedbrown
@souopgui There are points unchecked in your [review checklist](https://github.com/openjournals/joss-reviews/issues/1237#issue-406976574) at the top of this issue. Can you confirm those points so we can move forward with acceptance? Thanks!
