### 2019-04-06 #1237 @souopgui
@jedbrown I just checked those points based on the responses to my questions from the authors. I fully recommend the software.
