### 2019-04-06 #1237 @jedbrown
@grlee77 There is one citation `Stephan Mallat [@mallat2008wavelet]` that is "in-text" and should be written `Stephan Mallat [-@mallat2008wavelet]` or `@mallat2008wavelet`. See https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax for syntax. That's all I notice so I think we'll be ready to archive after that one minor fix.
