### 2019-04-06 74e7e17
DOC: JOSS: fix duplicate naming in in-text citation
