+++ doc/paper/paper.bib
@@ -10,7 +10,7 @@
                   Thomas Arildsen and
                   Helder Oliveira and
                   Ankit Agrawal and
-                  SylvainLan and
+                  Sylvain Lannuzel and
                   Michel Pelletier and
                   Matthew Brett and
                   Frank Yu and
@@ -34,7 +34,7 @@
   doi          = {10.5281/zenodo.1407172},
   url          = {https://doi.org/10.5281/zenodo.1407172}
 }
-TODO: determine real names for SylvainLan, 0-tree
+TODO: determine real name for 0-tree
 
 @Book{mallat2008wavelet,
   title = {A wavelet tour of signal processing: the sparse way},
@@ -129,7 +129,7 @@ month = {March},}
                   Johan Karlsson and
                   zickert and
                   Willem Jan Palenstijn and
-                  Ozan \"Oktem and
+                  \"Oktem Ozan and
                   Chong Chen and
                   Hector Andrade Loarca and
                   Michael Lohmann},
