### 2019-04-07 #1237 @grlee77
Thanks @jedbrown. It looks like the suggested `[-@mallat2008wavelet]` fixed the issue.
