### 2019-04-08 #1237 @jedbrown
Looks good, thanks. One more minor issue: please remove the last sentence (citing your old Zenodo archive). A link to the current Zenodo archive will be in the sidebar so the 2018 archive will just be confusing. After removing that sentence, go ahead and create a current archive and report the DOI here.
