### 2019-04-08 #1237 @grlee77
Okay, I can remove that sentence as well as the prior Zenodo reference from the .bib.

FYI, I was following the [example in the author guidelines](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) when I originally added that line. Perhaps the example should also remove that line if it is no longer recommended?

I think I will backport the `paper.md` and .bib to the 1.0.x branch and just tag a 1.0.3 release there. Does that sound fine? I don't think there are any new user-facing features to justify a 1.1 release from master (it has just dropped Python 2.x support and switched to `pytest` for testing).
