### 2019-04-08 f41b308
DOC: archive JOSS paper text
