+++ doc/paper/paper.md
@@ -1,0 +1,96 @@
+---
+title: 'PyWavelets: A Python package for wavelet analysis'
+tags:
+  - Python
+  - wavelets
+  - wavelet packets
+  - discrete wavelet transform
+  - continuous wavelet transform
+  - computational harmonic analysis
+authors:
+  - name: Gregory R. Lee
+    orcid: 0000-0001-8895-2740
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Ralf Gommers
+    orcid: 0000-0002-0300-3333
+    affiliation: "3, 4"
+  - name: Filip Waselewski
+    orcid: 0000-0003-0729-5879
+    affiliation:
+  - name: Kai Wohlfahrt
+    orcid: 0000-0002-0970-5539
+    affiliation: 5
+  - name: Aaron O'Leary
+    orcid: 0000-0003-1984-2323
+    affiliation:
+affiliations:
+ - name: Department of Radiology, Cincinnati Children's Hospital Medical Center, Cincinnati, OH, USA
+   index: 1
+ - name: Department of Radiology, University of Cincinnati School of Medicine, Cincinnati, OH, USA
+   index: 2
+ - name: Scion, 49 Sala Street, Private Bag 3020, Rotorua 3046, New Zealand
+   index: 3
+ - name: FPInnovations, 2665 East Mall, Vancouver, BC V6T 1Z4, Canada
+   index: 4
+ - name: Department of Biochemistry, University of Cambridge, Old Addenbrookes Site, 80 Tennis Court Road, Cambridge, CB2 1GA, United Kingdom
+   index: 5
+
+date: 24 August 2018
+bibliography: paper.bib
+---
+
+# Summary
+
+Wavelets are a popular tool for computational harmonic analysis. They provide
+localization in both the temporal (or spatial) domain as well as in the
+frequency domain [@Daubechies1992]. A prominent feature is the ability to
+perform a multiresolution analysis [@mallat2008wavelet]. The wavelet transform
+of natural signals and images tends to have most of its energy concentrated in
+a small fraction of the coefficients. This sparse representation property is
+key to the good performance of wavelets in applications such as data
+compression and denoising. For example, the wavelet transform is a key
+component of the JPEG 2000 image compression standard.
+
+``PyWavelets`` is a Python package implementing a number of n-dimensional
+discrete wavelet transforms as well as the 1D continuous wavelet transform. A
+wide variety of predefined wavelets are provided and it is possible for users
+to specify custom wavelet filter banks. All discrete wavelet transforms are
+implemented by convolution with finite impulse response filters. The required
+up/downsampling convolutions are implemented in C for good performance.
+Cython [@cython] is used to wrap the C code and implement axis-specific 1D
+transformations based on the low-level C routines. All multi-dimensional
+transforms are implemented in Python via separable application of the 1D
+transforms. The API for ``PyWavelets`` was designed to be similar to Matlab's
+wavelet toolbox and functions such as the 1D, 2D and 3D discrete wavelet
+transforms are tested for accuracy vs. their Matlab counterparts. PyWavelets
+has additional functionality not common in other wavelet toolboxes such as
+support for dimension $n > 3$ and support for both real and complex-valued
+data in either single or double precision. It is also possible to transform
+only a subset of axes and to vary the wavelet and signal boundary extension
+mode on a per-axis basis.
+
+``PyWavelets`` was designed for use by scientists working within a range of
+applications including time-series analysis, signal processing, image
+processing and medical imaging. It has already been adopted as a required
+or optional dependency by a number of other software projects. For example,
+it has enabled wavelet-based image denoising in scikit-image [@scikit-image].
+The Operator Discretization Library (ODL) [@odl] uses PyWavelets to enable
[…]
