+++ doc/paper/paper.bib
@@ -1,0 +1,106 @@
+
+
+@Book{mallat2008wavelet,
+  title = {A wavelet tour of signal processing: the sparse way},
+  author = {Mallat, Stéphane},
+  year = {2008},
+  publisher = {Academic press}
+}
+
+@book{Daubechies1992,
+ author = {Daubechies, Ingrid},
+ title = {Ten Lectures on Wavelets},
+ year = {1992},
+ isbn = {0-89871-274-2},
+ publisher = {Society for Industrial and Applied Mathematics},
+ address = {Philadelphia, PA, USA},
+ doi = {10.1137/1.9781611970104},
+ url = {https://doi.org/10.1137/1.9781611970104}
+}
+
+@Article{cython,
+author = {S. Behnel and R. Bradshaw and C. Citro and L. Dalcin and D. S. Seljebotn and K. Smith},
+journal = {Computing in Science Engineering},
+title = {Cython: The Best of Both Worlds},
+year = {2011},
+volume = {13},
+number = {2},
+pages = {31-39},
+keywords = {C language;numerical analysis;Python language extension;Fortran code;numerical loops;Cython language;programming language;Sparse matrices;Runtime;Syntactics;Computer programs;Programming;Python;Cython;numerics;scientific computing},
+doi = {10.1109/MCSE.2010.118},
+ISSN = {1521-9615},
+month = {March},}
+
+@Article{scikit-image,
+ title = {scikit-image: image processing in {P}ython},
+ author = {van der Walt, {S}t\'efan and {S}ch\"onberger, {J}ohannes {L}. and
+           {Nunez-Iglesias}, {J}uan and {B}oulogne, {F}ran\c{c}ois and {W}arner,
+           {J}oshua {D}. and {Y}ager, {N}eil and {G}ouillart, {E}mmanuelle and
+           {Y}u, {T}ony and the scikit-image contributors},
+ year = {2014},
+ month = {6},
+ keywords = {Image processing, Reproducible research, Education,
+             Visualization, Open source, Python, Scientific programming},
+ volume = {2},
+ pages = {e453},
+ journal = {PeerJ},
+ issn = {2167-8359},
+ url = {http://dx.doi.org/10.7717/peerj.453},
+ doi = {10.7717/peerj.453}
+}
+
+@article{kymatio,
+  author    = {Mathieu Andreux and
+               Tom{\'{a}}s Angles and
+               Georgios Exarchakis and
+               Roberto Leonarduzzi and
+               Gaspar Rochette and
+               Louis Thiry and
+               John Zarka and
+               St{\'{e}}phane Mallat and
+               Joakim And{\'{e}}n and
+               Eugene Belilovsky and
+               Joan Bruna and
+               Vincent Lostanlen and
+               Matthew J. Hirn and
+               Edouard Oyallon and
+               Sixhin Zhang and
+               Carmine{-}Emanuele Cella and
+               Michael Eickenberg},
+  title     = {Kymatio: Scattering Transforms in Python},
+  journal   = {CoRR},
+  volume    = {abs/1812.11214},
+  year      = {2018},
+  url       = {http://arxiv.org/abs/1812.11214},
+  archivePrefix = {arXiv},
+  eprint    = {1812.11214},
+  timestamp = {Wed, 02 Jan 2019 14:40:18 +0100},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1812-11214},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@Misc{odl,
+  author       = {Jonas Adler and
+                  Holger Kohr and
+                  Axel Ringh and
+                  Julian Moosmann and
+                  Sebastian Banert and
+                  Matthias J. Ehrhardt and
+                  Gregory R. Lee and
[…]
