+++ doc/source/releasenotes.rst
@@ -10,3 +10,4 @@ Release Notes
    release.1.0.0
    release.1.0.1
    release.1.0.2
+   release.1.0.3
