### 2019-04-08 #1237 @jedbrown
Thanks for pointing out the source of that verbiage. I've created a PR: https://github.com/openjournals/joss/pull/520
Yes, tagging v1.0.3 would be perfect.
