### 2019-04-09 d57fdfe
set version to 1.0.3 and ISRELEASED to True
