### 2019-04-09 #1237 @jedbrown
@grlee77 Could you please update the Zenodo archive's author list to match the paper?
