### 2019-04-09 #1237 @grlee77
There are six contributors who had an order of magnitude more commits and LOC than others and I was able to contact and get an affirmative response for authorship from 5 of the 6, so those are the ones I had listed for the JOSS publication. I was not able to get a response from Holger Nahrstaedt and so mentioned him and the other contributors in the acknowledgements instead.

Oddly, it seems that one of our top past contibutors, Kai Wohlfahrt (@kwohlfahrt) [no longer shows up as a contributor on the GitHub page](https://github.com/PyWavelets/pywt/graphs/contributors). I wonder why this is and if it could be a problem that occured when we recently had GitHub detach the repo from the depreceated github.com/nigma/PyWavelets repo? I assume this absence on GitHub is why he is also not present in the autogenerated Zenodo author list for v1.0.3.

I guess for now, I can manually edit the v1.0.3 release on Zenodo to match the author list here and add the appropriate ORCID IDs.
