### 2019-04-09 #1237 @grlee77
@rgommers: please see the prior comment on authorship. I'm curious if you have any idea why Kai's contributions are no longer visible?
