### 2019-04-09 #1237 @rgommers
No idea, that's very odd. If it's not a short glitch on GitHub's part, then it's worth opening a support issue about that I think.
