### 2019-04-09 #1237 @rgommers
> Could you please update the Zenodo archive's author list to match the paper?

It's not a big deal but for the record: I don't really like this. Paper author do not have to match software authors as a rule; there are authors who choose not to participate, or who can't be reached for permission to put their name on the paper. Manually removing one of our major contributors from Zenodo only for v1.0.3 doesn't feel quite right.
