### 2019-04-09 #1237 @jedbrown
:point_up: Cc: @openjournals/joss-eics
