### 2019-04-09 #1237 @grlee77
> Manually removing one of our major contributors from Zenodo only for v1.0.3 doesn't feel quite right.

I would also prefer to be more inclusive rather than less on Zenodo. @jedbrown, did you raise the issue because Kai wasn't listed there but was in the paper author list or does it need to be a 1:1 match?
