### 2019-04-09 #1237 @jedbrown
We normally ask for those lists to match. Note that JOSS reviews actually review the software, not just the paper. It sounds like there may be extenuating circumstances in this case. https://joss.readthedocs.io/en/latest/editing.html#after-acceptance
