### 2019-04-09 #1237 @rgommers
> It sounds like there may be extenuating circumstances in this case

okay thanks @jedbrown, then I suggest we add back our one major contributor who is not on the paper.
