### 2019-04-09 #1237 @rgommers
> Note that JOSS reviews actually review the software, not just the paper.

That doesn't really matter for who ends up being a paper author I would say. A JOSS paper is still a real academic paper with affiliations in it, so we can't just put people on that don't respond or don't want to be listed (I understand a match is desirable of course)
