### 2019-04-09 #1237 @grlee77
Pending any further feedback here, I added Holger (but not the full list of minor contributors) back to the author list for v1.0.3 on Zenodo for now. The first 5 authors listed there match the JOSS publication.
