### 2019-04-09 #1237 @labarba
> It's not a big deal but for the record: I don't really like this. Paper author do not have to match software authors as a rule; there are authors who choose not to participate, or who can't be reached for permission to put their name on the paper. Manually removing one of our major contributors from Zenodo only for v1.0.3 doesn't feel quite right.

JOSS assumes that the author list of the paper includes all major contributors. The paper is, after all, a _software paper_. As we state in the JOSS guidelines, it is up to the authors themselves to determine who should be credited with authorship. But you are right that authors should agree to be listed.

Zenodo grabs the author list automatically from all commits to the repo, and this usually bloats the list with folks who sent in a minor PR. It is rare (it's never come up before!) to have a major contributor who is not listed as an author in the paper.
