### 2019-04-09 #1237 @grlee77
@kwohlfahrt: If you would like to restore your commits to the contributors page again for PyWavelets, you will probably have to reach out to GitHub support. I got the following response from GitHub Support regarding their disappearance:

> Thanks for reaching out! In looking at an older commit made by that user in the PyWavelets/pywt repository, it appears that the email address used to author their commits is no longer linked to a GitHub account at all:
>
> https://github.com/PyWavelets/pywt/pull/59/commits/5831bbe89de12dfaa2385e71720cd4bc7fc196ed
>
> This is probably because the user removed that email from their account, and since GitHub can no longer 'link' the commits properly it will in turn mean they won't appear as a contributor to the repository. If you are in contact with the user, you are welcome to have them write in to us here in support and we can work with them to get those contributions linked back up.
