### 2019-04-09 #1237 @jedbrown
:point_up: Does GitHub not support `.mailmap` files (https://github.com/git/git/blob/master/.mailmap)?
