### 2019-04-10 #1237 @rgommers
> It is rare (it's never come up before!) to have a major contributor who is not listed as an author in the paper.

Interesting, thanks @labarba. It probably reflects that the vast majority of packages are young, modest in size and with few authors. For larger projects with a long history it's unlikely that all significant contributors are still active, can be reached and want to be listed. E.g. the SciPy paper we are finalizing right now will miss lots of names, it's unavoidable with a 15 year history and 700+ contributors. PyWavelets is not that large, but is 10+ years old.
