### 2019-04-10 #1237 @rgommers
> Does GitHub not support .mailmap files (https://github.com/git/git/blob/master/.mailmap)?

Good suggestion, worth a try at least. We have a manual name mapping for release notes that we can easily convert to `.mailmap`: https://github.com/PyWavelets/pywt/blob/master/util/authors.py#L27
