### 2019-04-10 #1237 @labarba
> For larger projects with a long history it's unlikely that all significant contributors are still active, can be reached and want to be listed.

In that case, a "best effort" approach should be made. Any contributors from long-ago who cannot be reached or who do not want to be authors can be listed in a long acknowledgement. The Zenodo author list should match the paper, as it is an archive deposit of the software associated with the software publication.
