### 2019-04-11 #1237 @grlee77
> In that case, a "best effort" approach should be made. Any contributors from long-ago who cannot be reached or who do not want to be authors can be listed in a long acknowledgement. The Zenodo author list should match the paper, as it is an archive deposit of the software associated with the software publication.

A "best effort" was already made for the top 6 authors (who account for ~95% of the total commits). I got an affirmative response from 5 of the 6 (~90% of total commits) and included these in the JOSS paper author list.

It sounds like you are insisting that we should only include the 5 JOSS paper authors for the 1.0.3 tag on Zenodo.

@rgommers: Are you okay with this? There should not be a restriction on going back to the full contributor list for future releases.

I did write in the release notes for 1.0.3 that it was archived specifically to correspond to this publication:
> "PyWavelets 1.0.3 is functionally equivalent to the 1.0.2 release. It was made to archive the JOSS paper about PyWavelets to the 1.0.x branch and serve as a reference corresponding to the version that was peer reviewed."

Perhaps a reasonable compromise is to edit that text on Zenodo to append an acknowledgement to all other contributrs to PyWavelets (potentially with a full list of names). That allows all names to be listed, just not as part of the official citation.
