### 2019-04-11 #1237 @rgommers
> @rgommers: Are you okay with this? There should not be a restriction on going back to the full contributor list for future releases.

As I said in my first post it's not a very big deal, so if the editors insist then I'm fine with it.

> Perhaps a reasonable compromise is to edit that text on Zenodo to append an acknowledgement to all other contributrs to PyWavelets (potentially with a full list of names). That allows all names to be listed, just not as part of the official citation.

That sounds like a good idea. And then for future releases we go back to the default Zenodo behavior.
