### 2019-04-11 #1237 @grlee77
>    Does GitHub not support .mailmap files (https://github.com/git/git/blob/master/.mailmap)?

We just added a `.mailmap` file in PyWavelets/pyw#485. This did clean up the output of `git shortlog` nicely, but does not appear to be used by GitHub itself.
