### 2019-04-11 #1237 @jedbrown
Good to know (if unfortunate) regarding `.mailmap`.
