### 2019-04-11 #1237 @grlee77
Please see the updated Zenodo entry. The author list there now matches the publication.
https://zenodo.org/record/2634243#.XK9g7UN7kQ8
