### 2019-04-11 #1237 @jedbrown
@labarba Do you know how to debug the crossref issue?
