### 2019-04-12 #1237 @arfon
Looks like Whedon doesn't like apostrophes in names (this feels like a new bug as I don't think I've seen it before). Short term fix here: https://github.com/PyWavelets/pywt/pull/487
